Removed local indexing
