+++ tests/DenseGraph.hpp
@@ -1,7 +1,7 @@
 #ifndef L3STER_DENSEGRAPH_HPP
 #define L3STER_DENSEGRAPH_HPP
 
-#include "l3ster/global_assembly/SparsityPattern.hpp"
+#include "l3ster/global_assembly/SparsityGraph.hpp"
 #include "l3ster/util/DynamicBitset.hpp"
 
 namespace lstr
@@ -16,7 +16,7 @@ public:
                const std::vector< global_dof_t >&                                          dofs)
         : dim{dofs.size()}, entries{dim * dim}
     {
-        const auto node_to_dof_map = GlobalNodeToDofMap{mesh, dof_intervals};
+        const auto node_to_dof_map = NodeToDofMap{mesh, dof_intervals};
         const auto process_domain  = [&]< auto dom_def >(ConstexprValue< dom_def >)
         {
             constexpr auto  domain_id        = dom_def.first;
